pyproject.toml
[project]
name = "edm-diffusers"
version = "0.1.0"
description = "EDM (Elucidated Diffusion Models) implemented with diffusers modules"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
  "torch>=2.2",
  "torchvision>=0.17",
  "diffusers>=0.27",
  "datasets>=2.16",
  "pyyaml>=6.0",
  "tqdm>=4.66",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
